requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1673535674859/work
aiohttp==3.8.4
aiosignal @ file:///home/conda/feedstock_root/build_artifacts/aiosignal_1667935791922/work
anyio==3.6.2
appdirs==1.4.4
async-timeout @ file:///home/conda/feedstock_root/build_artifacts/async-timeout_1640026696943/work
attrs==23.1.0
backoff==2.2.1
bleach==1.5.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1681349778161/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854167867/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1674482203741/work
certifi==2023.5.7
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==3.1.0
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
cmake==3.26.3
cog==0.6.1
contourpy==1.0.7
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1637687018854/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Deprecated==1.2.13
docker-pycreds==0.4.0
fastapi==0.95.1
filelock==3.11.0
flatbuffers==23.5.9
flit_core @ file:///croot/flit-core_1679397103445/work/source/flit_core
fonttools==4.39.3
frozenlist @ file:///croot/frozenlist_1670004507010/work
fsspec==2023.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth @ file:///home/conda/feedstock_root/build_artifacts/google-auth_1681369121356/work
google-auth-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/google-auth-oauthlib_1675825735130/work
googleapis-common-protos==1.59.0
grpcio==1.53.0
h11==0.14.0
h5py==3.8.0
html5lib==0.9999999
httptools==0.5.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1682141365603/work
Jinja2==3.1.2
joblib==1.2.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
libclang==16.0.0
lightning-utilities @ file:///croot/lightning-utilities_1677510309509/work
lit==16.0.1
lmdb==1.2.1
lpips==0.1.4
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1679584000376/work
MarkupSafe==2.1.2
matplotlib @ file:///croot/matplotlib-suite_1679593461707/work
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
netron==6.8.8
networkx==3.1
nibabel==5.1.0
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib @ file:///home/conda/feedstock_root/build_artifacts/oauthlib_1666056362788/work
onnx==1.14.0
opencv-python==4.7.0.72
opentelemetry-api==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-grpc==1.17.0
opentelemetry-exporter-otlp-proto-http==1.17.0
opentelemetry-proto==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
Pillow==9.5.0
ply==3.11
protobuf==4.22.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pyDeprecate==0.3.1
pydicom==2.3.1
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1666240235902/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytorch-fid==0.2.0
pytorch-lightning==1.4.5
pytz @ file:///croot/pytz_1671697431263/work
pyu2f @ file:///home/conda/feedstock_root/build_artifacts/pyu2f_1604248910016/work
PyYAML==6.0
redis==4.5.4
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1680286922386/work
requests-oauthlib @ file:///home/conda/feedstock_root/build_artifacts/requests-oauthlib_1643557462909/work
rsa @ file:///home/conda/feedstock_root/build_artifacts/rsa_1658328885051/work
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.19.1
setproctitle==1.3.2
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six==1.16.0
sklearn==0.0.post5
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1681422808449/work/tensorboard-2.12.2-py3-none-any.whl
tensorboard-data-server @ file:///croot/tensorboard-data-server_1681498183723/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl
tensorboard-plugin-wit @ file:///home/conda/feedstock_root/build_artifacts/tensorboard-plugin-wit_1641458951060/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-tensorboard==1.5.1
termcolor==2.3.0
threadpoolctl==3.1.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==2.0.1
torch-fidelity==0.3.0
torch-tb-profiler==0.4.1
torchmetrics==0.5.0
torchvision==0.15.2
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1678635778344/work
uvicorn==0.21.1
uvloop==0.17.0
wandb==0.14.2
watchfiles==0.19.0
websockets==11.0.1
Werkzeug==2.2.3
wrapt==1.14.1
yarl==1.8.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
